refactor folder structure to isolate managers -- now that classes reside within utils
